requirements.txt
Flask~=3.0.2
requests~=2.31.0
pillow~=10.2.0